extracted actual streaming response
