+++ nicegui/helpers.py
@@ -7,7 +7,12 @@ import threading
 import time
 import webbrowser
 from contextlib import nullcontext
+from pathlib import Path
 from typing import TYPE_CHECKING, Any, Awaitable, Callable, Optional, Tuple, Union
+
+import magic
+from fastapi import Request
+from fastapi.responses import StreamingResponse
 
 from . import background_tasks, globals
 
@@ -84,3 +89,38 @@ def schedule_browser(host: str, port: int) -> Tuple[threading.Thread, threading.
     thread = threading.Thread(target=in_thread, args=(host, port), daemon=True)
     thread.start()
     return thread, cancel
+
+
+def get_streaming_response(file: Path, request: Request) -> StreamingResponse:
+    file_size = file.stat().st_size
+    start, end = 0, file_size - 1
+    range_header = request.headers.get('Range', None)
+    if range_header:
+        byte1, byte2 = range_header.split('=')[1].split('-')
+        start = int(byte1)
+        if byte2:
+            end = int(byte2)
+    content_length = (end - start) + 1
+    headers = {
+        'Content-Range': f'bytes {start}-{end}/{file_size}',
+        'Content-Length': str(content_length),
+        'Accept-Ranges': 'bytes',
+    }
+
+    def content_reader(file, start, end, chunk_size=8192):
+        with open(file, 'rb') as data:
+            data.seek(start)
+            remaining_bytes = end - start + 1
+            while remaining_bytes > 0:
+                chunk = data.read(min(chunk_size, remaining_bytes))
+                if not chunk:
+                    break
+                yield chunk
+                remaining_bytes -= len(chunk)
+
+    return StreamingResponse(
+        content_reader(file, start, end),
+        media_type=magic.from_file(str(file), mime=True),
+        headers=headers,
+        status_code=206,
+    )
